LPS-121855 - Migrate management toolbar v2 usages in account-admin-web
