+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/display/context/AccountUsersAdminManagementToolbarDisplayContext.java
@@ -269,11 +269,6 @@ public class AccountUsersAdminManagementToolbarDisplayContext
 	}
 
 	@Override
-	public String getDefaultEventHandler() {
-		return "ACCOUNT_USERS_ADMIN_MANAGEMENT_TOOLBAR_DEFAULT_EVENT_HANDLER";
-	}
-
-	@Override
 	public List<DropdownItem> getFilterDropdownItems() {
 		List<DropdownItem> filterAccountEntriesDropdownItems =
 			_getFilterByAccountEntriesDropdownItems();
