+++ modules/apps/account/account-admin-web/src/main/resources/META-INF/resources/account_users_admin/select_account_entry.jsp
@@ -32,8 +32,8 @@ if (selectAccountEntryManagementToolbarDisplayContext.isSingleSelect()) {
 }
 %>
 
-<clay:management-toolbar-v2
-	displayContext="<%= selectAccountEntryManagementToolbarDisplayContext %>"
+<clay:management-toolbar
+	managementToolbarDisplayContext="<%= selectAccountEntryManagementToolbarDisplayContext %>"
 />
 
 <clay:container-fluid
